Renamed benchmarked weight files
